Add more components-- not sure what the current problem is -- says base class undefined

+++ Engine/Private/Chunk.cpp
@@ -45,7 +45,7 @@ GLvoid Chunk::initChunks(GLuint programIn, mat4* viewMatIn, const uvec2& chunksS
 
 }
 
-Chunk::Chunk(vec2 locationIn)
+Chunk::Chunk(ivec2 locationIn)
 	: location(locationIn)
 {
